requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
aiohttp=3.7.3=pypi_0
argon2-cffi=20.1.0=py39hbd71b63_2
async-timeout=3.0.1=pypi_0
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=3.3.0=pyh44b312d_0
blosc=1.20.1=hd408876_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h38d8fee_1001
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.4.13=h06a4308_1
cachetools=4.2.1=pypi_0
certifi=2020.12.5=py39h06a4308_0
cffi=1.14.4=py39h261ae71_0
chardet=3.0.4=pypi_0
charls=2.1.0=he6710b0_2
click=7.1.2=pypi_0
cloudpickle=1.6.0=py_0
configparser=5.0.2=pypi_0
cryptography=3.3.1=py39h3c74f83_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py39h06a4308_0
cytoolz=0.11.0=py39h27cfd23_0
dask-core=2021.1.1=pyhd3eb1b0_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
et_xmlfile=1.0.1=py_1001
freetype=2.10.4=h5ab3b9f_0
fsspec=0.8.5=pypi_0
future=0.18.2=pypi_0
giflib=5.1.4=h14c3975_1
gitdb=4.0.7=pypi_0
gitpython=3.1.14=pypi_0
google-auth=1.26.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
grpcio=1.35.0=pypi_0
h5py=2.10.0=py39hec9cf62_0
hdf5=1.10.6=hb1b8bf9_0
idna=2.10=pyh9f0ad1d_0
imagecodecs=2021.1.11=py39h581e88b_1
imageio=2.9.0=py_0
importlib-metadata=4.0.1=py39hf3d152e_0
intel-openmp=2020.2=254
ipykernel=5.5.3=py39hef51801_0
ipython=7.22.0=py39hef51801_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3deb0d_0
jdcal=1.4.1=py_0
jedi=0.18.0=py39hf3d152e_2
jinja2=2.11.3=pyh44b312d_0
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py39hf3d152e_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py39h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
lerc=2.2.1=h2531618_0
libaec=1.0.4=he6710b0_1
libdeflate=1.7=h27cfd23_5
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libopenblas=0.3.10=h5a2b251_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libwebp=1.0.1=h8e7db2f_0
libzopfli=1.0.3=he6710b0_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.3=pypi_0
markupsafe=1.1.1=py39h38d8fee_2
matplotlib-base=3.3.2=py39hd261b2b_0
mistune=0.8.4=py39hbd71b63_1002
mkl=2020.2=256
mkl-service=2.3.0=py39he8ac12f_0
mkl_fft=1.0.6=py39h63df603_0
mkl_random=1.0.2=py39h63df603_0
multidict=5.1.0=pypi_0
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py39hf3d152e_3
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
networkx=2.5=py_0
ninja=1.10.2=py39hff7bd54_0
nose=1.3.7=py_1006
notebook=6.3.0=pyha770c72_1
numpy=1.19.2=py39h89c1606_0
numpy-base=1.19.2=py39h2ae0177_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openjpeg=2.3.0=h05c96fa_1
openpyxl=3.0.7=pyhd3eb1b0_0
openssl=1.1.1k=h27cfd23_0
packaging=20.9=pyh44b312d_0
pandas=1.2.1=py39ha9443f7_0
pandoc=2.12=h7f98852_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd8ed1ab_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.1.0=py39he98fc37_0
pip=20.3.3=py39h06a4308_0
prometheus_client=0.10.1=pyhd8ed1ab_0
promise=2.3=pypi_0
prompt-toolkit=3.0.18=pyha770c72_0
protobuf=3.14.0=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.8.1=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.17.3=py39hbd71b63_1
pysocks=1.7.1=py39hf3d152e_3
python=3.9.1=hdb3f193_2
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.9=1_cp39
pytorch=1.7.1=py3.9_cuda11.0.221_cudnn8.0.5_0
pytorch-lightning=1.1.8=pypi_0
pytorch-lightning-bolts=0.3.0=pypi_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py39h6323ea4_4
pyyaml=5.3.1=py39h38d8fee_1
pyzmq=19.0.2=py39hb69f2a1_2
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7=pypi_0
scikit-image=0.18.1=py39ha9443f7_0
scikit-learn=0.23.2=py39ha9443f7_0
scipy=1.6.0=py39h91f5cce_0
send2trash=1.5.0=py_0
sentry-sdk=1.0.0=pypi_0
setuptools=52.0.0=py39h06a4308_0
shortuuid=1.0.1=pypi_0
six=1.15.0=py39h06a4308_0
smmap=4.0.0=pypi_0
snappy=1.1.8=he6710b0_0
sqlite=3.33.0=h62c20be_0
subprocess32=3.5.4=pypi_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
terminado=0.9.4=py39hf3d152e_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.1.14=pyhd3eb1b0_1
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchvision=0.8.2=py39_cu110
tornado=6.1=py39h27cfd23_0
tqdm=4.56.2=pypi_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=pyha847dfd_0
tzdata=2020f=h52ac0ba_0
urllib3=1.26.3=pyhd8ed1ab_0
wandb=0.10.27=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=1.0.1=pypi_0
wfdb=3.2.0=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py39hf3d152e_4
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=pypi_0
zeromq=4.3.4=h2531618_0
zfp=0.5.5=h2531618_4
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
